requirements.txt
beautifulsoup4
colorama
imdbpy
imageio
prettytable
python-dotenv
requests